Move contents of starfish/register.py to starfish/registration/_fourier_shift.py
